test/console.test.ts
import { promptForInput } from '../src/console';


describe("console test", () => {
    const mockCallback = jest.fn((str : string) => str.toUpperCase());
    test("input test", () => {  
        /*promptForInput("", mockCallback );  
        promptForInput("hello", mockCallback );  
        expect(mockCallback.mock.calls).toHaveLength(2);
        expect(mockCallback.mock.calls[0][0]).toBe("");
        expect(mockCallback.mock.results[0].value).toBe("");
        expect(mockCallback.mock.calls[0][0]).toBe("hello");
        expect(mockCallback.mock.results[0].value).toBe("HELLO");*/

        expect(mockCallback.mock.calls).toHaveLength(0);    // dummy test
    });
    // todo not working yet, trying to check the callback is called
    // that is passed to promptForInput, need to wait or console input
    // so not easy to test
});